Create OverlayedStorageEntry with its own specific implementation
