+++ internal/primitives/state-machine/overlayedchanges/overlayed_map.go
@@ -19,9 +19,9 @@ var errorNotInRuntime = errors.New("not in runtime")
 var errorNoOpenTransaction = errors.New("no open transaction")
 
 // Holds a set of changes with the ability modify them using nested transactions.
-type OverlayedMap[K constraints.Ordered, V any] struct {
+type OverlayedMap[K constraints.Ordered, V any, E OverlayedEntry[V]] struct {
 	// Stores the changes that this overlay constitutes.
-	changes btree.Map[K, *OverlayedEntry[V]]
+	changes btree.Map[K, E]
 	// Stores which keys are dirty per transaction. Needed in order to determine which
 	// values to merge into the parent transaction on commit. The length of this vector
 	// therefore determines how many nested transactions are currently open (depth).
@@ -34,8 +34,8 @@ type OverlayedMap[K constraints.Ordered, V any] struct {
 	executionMode executionMode
 }
 
-func NewOverlayedMap[K constraints.Ordered, V any]() OverlayedMap[K, V] {
-	return OverlayedMap[K, V]{
+func NewOverlayedMap[K constraints.Ordered, V any, E OverlayedEntry[V]]() OverlayedMap[K, V, E] {
+	return OverlayedMap[K, V, E]{
 		dirtyKeys:             dirtyKeysSets[K]{},
 		numClientTransactions: 0,
 		executionMode:         executionModeClient,
@@ -45,8 +45,8 @@ func NewOverlayedMap[K constraints.Ordered, V any]() OverlayedMap[K, V] {
 // Create a new changeset at the same transaction state but without any contents.
 // This changeset might be created when there are already open transactions.
 // We need to catch up here so that the child is at the same transaction depth.
-func (om *OverlayedMap[K, V]) SpawnChild() OverlayedMap[K, V] {
-	return OverlayedMap[K, V]{
+func (om *OverlayedMap[K, V, E]) SpawnChild() OverlayedMap[K, V, E] {
+	return OverlayedMap[K, V, E]{
 		dirtyKeys:             make(dirtyKeysSets[K], om.TransactionDepth()),
 		numClientTransactions: om.numClientTransactions,
 		executionMode:         om.executionMode,
@@ -54,48 +54,43 @@ func (om *OverlayedMap[K, V]) SpawnChild() OverlayedMap[K, V] {
 }
 
 // True if no changes at all are contained in the change set.
-func (om *OverlayedMap[K, V]) IsEmpty() bool {
+func (om *OverlayedMap[K, V, E]) IsEmpty() bool {
 	return om.changes.Len() == 0
 }
 
 // Get an optional reference to the value stored for the specified key.
-func (om *OverlayedMap[K, V]) Get(key K) *OverlayedEntry[V] {
-	value, present := om.changes.Get(key)
-	if !present {
-		return nil
-	}
-
-	return value
+func (om *OverlayedMap[K, V, E]) Get(key K) (E, bool) {
+	return om.changes.Get(key)
 }
 
 // Set a new value for the specified key.
 // Can be rolled back or committed when called inside a transaction.
-func (om *OverlayedMap[K, V]) SetOffchain(key K, value V, atExtrinsic *uint32) {
-	overlayed := om.Get(key)
+func (om *OverlayedMap[K, V, E]) SetOffchain(key K, value V, atExtrinsic *uint32) {
+	overlayed, has := om.Get(key)
 
-	if overlayed == nil {
-		overlayed = NewOverlayedEntry[V]()
+	if !has {
+		overlayed = *new(E)
 	}
 
 	overlayed.SetOffchain(value, om.dirtyKeys.insertDirty(key), atExtrinsic)
 }
 
 // Get a list of all changes as seen by current transaction.
-func (om *OverlayedMap[K, V]) Changes() iter.Seq2[K, *OverlayedEntry[V]] {
-	return func(yield func(K, *OverlayedEntry[V]) bool) {
+func (om *OverlayedMap[K, V, E]) Changes() iter.Seq2[K, E] {
+	return func(yield func(K, E) bool) {
 		om.changes.Scan(yield)
 	}
 }
 
 // Return all committed changes.
 // Panics if there are open transactions: `transaction_depth() > 0`
-func (om *OverlayedMap[K, V]) DrainCommited() iter.Seq2[K, V] {
+func (om *OverlayedMap[K, V, E]) DrainCommited() iter.Seq2[K, V] {
 	if om.TransactionDepth() != 0 {
 		panic("Drain is not allowed with open transactions.")
 	}
 
 	return func(yield func(K, V) bool) {
-		om.changes.Scan(func(k K, v *OverlayedEntry[V]) bool {
+		om.changes.Scan(func(k K, v E) bool {
 			return yield(k, v.PopTransaction().value)
 		})
 	}
@@ -103,14 +98,14 @@ func (om *OverlayedMap[K, V]) DrainCommited() iter.Seq2[K, V] {
 
 // Returns the current nesting depth of the transaction stack.
 // A value of zero means that no transaction is open and changes are committed on write.
-func (om *OverlayedMap[K, V]) TransactionDepth() uint {
+func (om *OverlayedMap[K, V, E]) TransactionDepth() uint {
 	return uint(len(om.dirtyKeys))
 }
 
 // Call this before transferring control to the runtime.
 // This protects all existing transactions from being removed by the runtime.
 // Calling this while already inside the runtime will return an error.
-func (om *OverlayedMap[K, V]) EnterRuntime() error {
+func (om *OverlayedMap[K, V, E]) EnterRuntime() error {
 	if om.executionMode == executionModeRuntime {
 		return errorAlreadyInRuntime
 	}
@@ -123,7 +118,7 @@ func (om *OverlayedMap[K, V]) EnterRuntime() error {
 // Call this when control returns from the runtime.
 // This rollbacks all dangling transaction left open by the runtime.
 // Calling this while already outside the runtime will return an error.
-func (om *OverlayedMap[K, V]) ExitRuntimeoffchain() error {
+func (om *OverlayedMap[K, V, E]) ExitRuntimeoffchain() error {
 	if om.executionMode != executionModeRuntime {
 		return errorNotInRuntime
 	}
@@ -150,26 +145,26 @@ func (om *OverlayedMap[K, V]) ExitRuntimeoffchain() error {
 // transaction was open. Any transaction must be closed by either `commit_transaction`
 // or `rollback_transaction` before this overlay can be converted into storage changes.
 // Changes made without any open transaction are committed immediately.
-func (om *OverlayedMap[K, V]) StartTransaction() {
+func (om *OverlayedMap[K, V, E]) StartTransaction() {
 	om.dirtyKeys = append(om.dirtyKeys, map[K]struct{}{})
 }
 
 // Rollback the last transaction started by `start_transaction`.
 // Any changes made during that transaction are discarded. Returns an error if
 // there is no open transaction that can be rolled back.
-func (om *OverlayedMap[K, V]) RollbackTransactionOffchain() error {
+func (om *OverlayedMap[K, V, E]) RollbackTransactionOffchain() error {
 	return om.closeTransactionOffchain(true)
 }
 
 // Commit the last transaction started by `start_transaction`.
 // Any changes made during that transaction are committed. Returns an error if
 // there is no open transaction that can be committed.
-func (om *OverlayedMap[K, V]) CommitTransactionOffchain() error {
+func (om *OverlayedMap[K, V, E]) CommitTransactionOffchain() error {
 	return om.closeTransactionOffchain(false)
 }
 
 // Internal method to close the transaction and either commit or roll back the changes.
-func (om *OverlayedMap[K, V]) closeTransactionOffchain(rollback bool) error {
+func (om *OverlayedMap[K, V, E]) closeTransactionOffchain(rollback bool) error {
 	// runtime is not allowed to close transactions started by the client
 	if om.executionMode == executionModeRuntime && !om.HasOpenRuntimeTransactions() {
 		return errorNoOpenTransaction
@@ -195,7 +190,7 @@ func (om *OverlayedMap[K, V]) closeTransactionOffchain(rollback bool) error {
 
 			// We need to remove the key as an `OverlayValue` with no transactions
 			// violates its invariant of always having at least one transaction.
-			if len(overlayed.transactions) == 0 {
+			if len(overlayed.Transactions()) == 0 {
 				om.changes.Delete(key)
 			}
 		} else {
@@ -211,7 +206,7 @@ func (om *OverlayedMap[K, V]) closeTransactionOffchain(rollback bool) error {
 				// Last tx: Is there already a value in the committed set?
 				// Check against one rather than empty because the current tx is still
 				// in the list as it is popped later in this function.
-				hasPredecessor = len(overlayed.transactions) > 1
+				hasPredecessor = len(overlayed.Transactions()) > 1
 			}
 
 			// We only need to merge if there is an pre-existing value. It may be a value from
@@ -228,6 +223,6 @@ func (om *OverlayedMap[K, V]) closeTransactionOffchain(rollback bool) error {
 }
 
 // True if there are open transactions that were started by the runtime.
-func (om *OverlayedMap[K, V]) HasOpenRuntimeTransactions() bool {
+func (om *OverlayedMap[K, V, E]) HasOpenRuntimeTransactions() bool {
 	return om.TransactionDepth() > om.numClientTransactions
 }
